Add option to pgfplot compatibility
@abelsiqueira request low the value of pgfplot compatibility option.
Low the value is not the best solution due it is dependent of the
installation.

To solve the problem we add a new option in the command line. If the
user don't use it perprof-py will be compatible with the newer version.

A brief explanation about this solution can be found at
http://tex.stackexchange.com/a/81912/22451.